+++ packages/backend/src/nest/common/test.module.ts
@@ -25,12 +25,10 @@ import { createTmpDir, getCors, torBinForPlatform, torDirForPlatform } from './u
 
 const torPath = torBinForPlatform()
 const libPath = torDirForPlatform()
-// torBinaryPath: '../../../../../3rd-party/tor/linux/tor',
-// torResourcesPath: '../../../../../3rd-party/tor/linux',
 export const defaultConfigForTest = {
   socketIOPort: TEST_DATA_PORT,
   torBinaryPath: torBinForPlatform(),
-  torResourcesPath: torPath,
+  torResourcesPath: torDirForPlatform(),
   torControlPort: await getPort(),
   options: {
     env: {