+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+[unreleased]
+
+* Add community metadata validation.
+
+* Move community metadata to separate store.
+
 [2.0.3-alpha.13]
 
 * Initialize electron-store after setting appData to prevent creating empty "Quiet" data directory